requirements.txt
google-generativeai
pandas
pydantic
requests